Refactor to async/await

+++ server/server.js
@@ -22,87 +22,84 @@ app.get('/users/me', authenticate, (req, res) => {
     res.send(req.user)
 })
 
-app.post('/users/login', (req, res) => {
-    let body = _.pick(req.body, ['email', 'password'])
-
-    User.findByCredentials(body.email, body.password).then((user) => {
-        user.generateAuthToken().then((token) => {
-            res.header('x-auth', token).send(user)
-        })
-    }).catch((e) => {
-        res.status(400).send()
-    })
-})
-
-app.delete('/users/me/token', authenticate, (req, res) => {
-    req.user.removeToken(req.token).then(() => {
-        res.status(200).send()
-    })
-})
-
-app.post('/users', (req, res) => {
-    let body = _.pick(req.body, ['email', 'password'])
-    let user = new User(body)
-
-    user.save().then((user) => {
-        return user.generateAuthToken()
-    }).then((token) => {
+app.post('/users/login', async (req, res) => {
+    try {
+        const body = _.pick(req.body, ['email', 'password'])
+        const user = await User.findByCredentials(body.email, body.password)
+        const token = await user.generateAuthToken()
         res.header('x-auth', token).send(user)
-    }).catch((e) => {
+    } catch(e) {
+        res.status(400).send()
+    }
+})
+
+app.delete('/users/me/token', authenticate, async (req, res) => {
+    try {
+        await req.user.removeToken(req.token)
+        res.status(200).send()
+    } catch(e) {
+        res.status(400).send()
+    }
+})
+
+app.post('/users', async (req, res) => {
+    const body = _.pick(req.body, ['email', 'password'])
+    try {
+        const user = new User(body)
+        await user.save()
+        const token = await user.generateAuthToken()
+        res.header('x-auth', token).send(user)
+    } catch (e) {
         console.log("Error", e)
         res.status(400).send()
-    })
+    }
 })
 
-app.get('/todos', authenticate, (req, res) => {
-    Todo.find({
-        _creator: req.user.id
-    }).then((todos) => {
-        res.send({todos})
-    }).catch((e) => {
-        console.log("Error")
-        res.status(400).send(e)
-    })
+app.get('/todos', authenticate, async (req, res) => {
+    try {
+        const todos = await Todo.find({_creator: req.user.id})
+        res.send({ todos })
+    } catch (e) {
+        console.log(e)
+        res.status(400).send()
+    }
 })
 
-app.get('/todos/:id', authenticate, (req, res) => {
-    let id = req.params.id
+app.get('/todos/:id', authenticate, async (req, res) => {
+    const id = req.params.id
     if(!ObjectID.isValid(id)) {
         return res.status(404).send()
     }
 
-    Todo.findOne({
-            _id: id,
-            _creator: req.user._id
-        }
-    ).then((todo) => {
+    try {
+        const todo = await Todo.findOne({_id: id, _creator: req.user._id})
         if(!todo) {
-            return res.status(404).send()
+            throw new Error(`Item ${id} cannot be found in your collection`)
         }
-
-        res.send({todo})
-    }).catch((e) => {
-        console.log("Error")
+        res.send(todo)
+    } catch (e) {
+        console.log(e)
         res.status(400).send()
-    })
+    }
 })
 
-app.post('/todos', authenticate, (req, res) => {
-    let todo = new Todo({
+app.post('/todos', authenticate, async (req, res) => {
+    const todo = new Todo({
         text: req.body.text,
         _creator: req.user._id
     })
 
-    todo.save().then((doc) => {
+    try {
+        const doc = await todo.save()
         res.send(doc)
-    }).catch((e) => {
-        console.log("Error")
+    } catch (e) {
+        console.log(e)
         res.status(400).send()
-    })
+    }
 })
 
-app.patch('/todos/:id', authenticate, (req, res) => {
-    let id = req.params.id
+app.patch('/todos/:id', authenticate, async (req, res) => {
+    const id = req.params.id
     let body = _.pick(req.body, ['text', 'completed'])
     
     if (!ObjectID.isValid(id)) {
@@ -116,36 +113,34 @@ app.patch('/todos/:id', authenticate, (req, res) => {
         body.completedAt = null
     }
 
-    Todo.findOneAndUpdate({
-        _id: id,
-        _creator: req.user._id
-    }, {$set: body}, {new: true}).then((todo) => {
+    try {
+        const todo = await Todo.findOneAndUpdate({
+            _id: id,
+            _creator: req.user._id
+        }, { $set: body }, { new: true })
         res.send(todo)
-    }).catch((e) => {
+    } catch (e) {
+        console.log(e);
         res.status(400).send()
-    })
+    }
 })
 
-app.delete('/todos/:id', authenticate, (req, res) => {
-    let id = req.params.id
+app.delete('/todos/:id', authenticate, async (req, res) => {
+    const id = req.params.id
     if (!ObjectID.isValid(id)) {
         return res.status(404).send()
     }
 
-    Todo.findOneAndRemove({
-            _id: id,
-            _creator: req.user.id
+    try {
+        const todo = await Todo.findOneAndRemove({_id: id, _creator: req.user.id})
+        if(!todo) {
+            throw new Error(`Unable to delete ${id}.`)
         }
-    ).then((todo) => {
-        if (!todo) {
-            return res.status(404).send()
-        }
-
         res.send({ todo })
-    }).catch((e) => {
-        console.log("Error")
+    } catch (e) {
+        console.log(e)
         res.status(400).send()
-    })
+    }
 })
 
 app.listen(port, () => {
